Reverting Coredump status field in Reboot Info struct
Though we have commandeered the use of the cordump_saved field to track failures in the coredump save routine, it is still expected by memfault that this variable will be a bool reflecting the true/false status of whether a coredump is present. As such, I have changed the name of the hijacked variable back to it's original name, but it is still doing the dual duty of tracking coredump failures for us, and indicating coredump present for memfault. Upon reset, if a coredump_saved status code indicates a save failure, then this is captured and coredump_saved is reassigned to false to stay compatible. If a dump saved successfully, then it is already holding a value of true.